+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/ProcessActionRepository.scala
@@ -146,6 +146,18 @@ extends Repository[F] with EspTables with CommentActions with ProcessActionRepos
     run(query.result.map(_.toSet))
   }
 
+  def getLastStateActionType(processId: ProcessId): F[Option[ProcessActionEntityData]] = {
+    val query = processActionsTable
+      .filter(action =>
+        action.processId === processId &&
+        action.action.inSet(ProcessActionType.StateActions) &&
+        action.state === ProcessActionState.Finished
+      )
+      .sortBy(_.performedAt.desc)
+
+    run(query.result.headOption)
+  }
+
   def getUserActionsAfter(user: LoggedUser, possibleActionTypes: Set[ProcessActionType], possibleStates: Set[ProcessActionState], limit: Instant): F[Seq[(ProcessActionEntityData, ProcessName)]] = {
     run(
       processActionsTable