+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/api/deployment/ProcessAction.scala
@@ -16,6 +16,8 @@ import java.time.Instant
                                     buildInfo: Map[String, String]) {
   def isDeployed: Boolean = action.equals(ProcessActionType.Deploy)
   def isCanceled: Boolean = action.equals(ProcessActionType.Cancel)
+  def isArchived: Boolean = action.equals(ProcessActionType.Archive)
+  def isUnArchived: Boolean = action.equals(ProcessActionType.UnArchive)
 }
 
 object ProcessActionType extends Enumeration {
@@ -29,7 +31,8 @@ object ProcessActionType extends Enumeration {
   val UnArchive: Value = Value("UNARCHIVE")
   val Pause: Value = Value("PAUSE") //TODO: To implement in future..
 
-  val defaultActions: List[ProcessActionType] = Nil
+  val DefaultActions: List[ProcessActionType] = Nil
+  val StateActions = List(Cancel, Deploy)
 }
 
 object ProcessActionState extends Enumeration {