+++ docs/Changelog.md
@@ -37,6 +37,7 @@
 * [#4332](https://github.com/TouK/nussknacker/pull/4332) Improvements: Don't fetch state for fragments at /api/processes/status
 * [#4326](https://github.com/TouK/nussknacker/pull/4326) Expressions validation mechanism now is accumulating as many errors as possible instead of failing fast - for purpose 
   of backend code completion and dictionaries substitution
+* [#4339](https://github.com/TouK/nussknacker/pull/4339) Improvements: Don't fetch state for archived/unarchived scenario, return computed based on last state action
 
 1.9.1 (24 Apr 2023)
 ------------------------